fix style import & dashboard routes
