+++ app/(dashboard)/dashboard/getstarted/config.ts
