+++ app/(dashboard)/dashboard/onboarding/config.ts
