+++ app/(dashboard)/dashboard/task-list/page.tsx
@@ -4,7 +4,7 @@ import { QuestProvider, TutorialScreen } from "@questlabs/react-sdk"
 
 import { apikey, entityId, questId, token, userId } from "./config"
 import "@questlabs/react-sdk/dist/style.css"
-import "../../styles/quest.css"
+import "../../../../styles/quest.css"
 import { DashboardHeader } from "@/components/header"
 import { DashboardShell } from "@/components/shell"
